delete from cart action, minor jsdocs fixes
