+++ _dev/js/theme/core/cart/index.js
@@ -1,6 +1,5 @@
 import updateCart from '@js/theme/core/cart/updateCart';
 import cartQuantity from '@js/theme/core/cart/cartQuantity';
-import cartDelete from '@js/theme/core/cart/cartDelete';
 import DOMReady from '@js/theme/utils/DOMReady';
 import cartController from './cartController';
 
@@ -9,6 +8,5 @@ const { init } = cartController();
 DOMReady(() => {
   updateCart();
   cartQuantity();
-  cartDelete();
   init();
 });
