+++ _dev/js/theme/core/checkout/utils/switchEditAddressButtonColor.js
@@ -2,9 +2,9 @@ import { each } from '../../../utils/DOMHelpers';
 
 /**
  * Change the color of the edit button for the wrong address
- * @param {Boolean} enabled - true if button should be dangered or false otherwise
- * @param {Number} id - address id
- * @param {String} type - address type (delivery or invoice)
+ * @param {boolean} enabled - true if button should be dangered or false otherwise
+ * @param {int} id - address id
+ * @param {string} type - address type (delivery or invoice)
  */
 const switchEditAddressButtonColor = (
   enabled,
