+++ _dev/js/theme/frontAPI/apiAction.js
@@ -1,6 +1,5 @@
 import refreshCartPageAction from '@js/theme/frontAPI/cart/refreshCartPageAction';
 import updateCartQuantityAction from '@js/theme/frontAPI/cart/updateCartQuantityAction';
-import deleteFromCartAction from '@js/theme/frontAPI/cart/deleteFromCartAction';
 import updateProductAction from '@js/theme/frontAPI/product/updateProductAction';
 import updateListingFacetsAction from '@js/theme/frontAPI/listing/updateListingFacetsAction';
 import updateAddressAction from '@js/theme/frontAPI/address/updateAddressAction';
@@ -17,7 +16,6 @@ prestashop.addAction = (actionName, actionFunction) => {
 
 prestashop.addAction('refreshCartPage', refreshCartPageAction);
 prestashop.addAction('updateCartQuantity', updateCartQuantityAction);
-prestashop.addAction('deleteFromCart', deleteFromCartAction);
 prestashop.addAction('updateProduct', updateProductAction);
 prestashop.addAction('updateListingFacets', updateListingFacetsAction);
 prestashop.addAction('updateAddress', updateAddressAction);
